\name{set120}
\alias{set120}
\docType{data}

\title{
Dataset containing 120 values
}

\description{
Small dataset containing 120 samples.  The mean is centered on 27 with a standard deviation
of 7.
}

\usage{set120}
\format{
  The format is:
 num [1:120] 38.1 12.6 31.3 35.5 22.6 ...
}

\source{
\code{rnorm(120, m = 27, sd = 7)}
}

\examples{
data(set120)
}

\keyword{datasets}
