\name{set20}
\alias{set20}
\docType{data}

\title{
Small dataset containing 20 samples
}

\description{
Small dataset containing 20 samples.  The mean is centered on 42 with a standard deviation
of 5.
}

\usage{set20}
\format{
  The format is:
 num [1:20] 35 32.9 43.6 44.6 35.9 ...
}

\source{
\code{rnorm(20, m = 42, sd = 6)}
}

\examples{
data(set20)
}

\keyword{datasets}
