\name{set200}
\alias{set200}
\docType{data}

\title{
Dataset containing 200 values
}

\description{
Small dataset containing 200 samples.  The mean is centered on 5 with a standard deviation
of 1.
}

\usage{set200}
\format{
  The format is:
 num [1:200] 3.95 5.16 5.32 3.86 3.54 ...
}

\source{
\code{rnorm(200, m = 5, sd = 1)}
}

\examples{
data(set200)
}

\keyword{datasets}
