\name{set50}
\alias{set50}
\docType{data}

\title{
Dataset containing 50 values
}

\description{
Small dataset containing 50 samples.  The mean is centered on 14 with a standard deviation
of 3.
}

\usage{set50}
\format{
  The format is:
 num [1:50] 16.61 20.43 7.91 15.19 14.77 ...
}

\source{
\code{rnorm(50, m = 14, sd = 3)}
}

\examples{
data(set50)
}

\keyword{datasets}
