\name{tables.RData}
\alias{tables.RData}
\docType{data}

\title{
Two reference tables necessary for reference interval calculations
}

\description{
Table for non-parametric ranks of confidence intervals.
}

\usage{tables.RData}
\format{
  The format is:
 chr "tables.RData"
}

\examples{
data(tables.RData)
}

\keyword{datasets}
